# Improve dissect performance