<!-- Type of change
Please label this PR with one of the following labels, depending on the scope of your change:
- Bug
- Enhancement
- Breaking change
- Deprecation
- Cleanup
- Docs
-->

## What does this PR do?

Improve dissect performance, especially when there is data type conversion defined in dissect pattern. 

## Why is it important?

performance improvement.

## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to have the PR approved

List here all the items you have verified BEFORE sending this PR. Please DO NOT remove any item, striking through those that do not apply. (Just in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] ~~I have made corresponding changes to the documentation~~
- [x] ~~I have made corresponding change to the default configuration files~~
- [x] ~~I have added tests that prove my fix is effective or that my feature works~~
- [x] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.

## Author's Checklist

<!-- Recommended
Add a checklist of things that are required to be reviewed in order to have the PR approved
-->
- [ ] test dissect package

## How to test this PR locally

this can be simply tested by go test the dissect package.

## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
- 

## Use cases

<!-- Recommended
Explain here the different behaviors that this PR introduces or modifies in this project, user roles, environment configuration, etc.

If you are familiar with Gherkin test scenarios, we recommend its usage: https://cucumber.io/docs/gherkin/reference/
-->

## Screenshots

<!-- Optional
Add here screenshots about how the project will be changed after the PR is applied. They could be related to web pages, terminal, etc, or any other image you consider important to be shared with the team.
-->

## Logs

<!-- Recommended
Paste here output logs discovered while creating this PR, such as stack traces or integration logs, or any other output you consider important to be shared with the team.
-->
```
name                                   old time/op    new time/op    delta
DissectNoConversionOneValue-8            7.68µs ±39%    6.27µs ± 5%  -18.44%  (p=0.000 n=20+19)
DissectWithConversionOneValue-8          7.44µs ±10%    6.78µs ± 9%   -8.91%  (p=0.000 n=19+19)
DissectNoConversionMultipleValues-8      19.7µs ± 9%    18.5µs ± 5%   -6.27%  (p=0.000 n=19+18)
DissectWithConversionMultipleValues-8    22.9µs ± 7%    21.0µs ± 6%   -8.42%  (p=0.000 n=18+19)

name                                   old alloc/op   new alloc/op   delta
DissectNoConversionOneValue-8            2.64kB ± 0%    2.64kB ± 0%     ~     (p=0.264 n=19+20)
DissectWithConversionOneValue-8          2.64kB ± 0%    2.64kB ± 0%     ~     (p=0.413 n=20+20)
DissectNoConversionMultipleValues-8      5.81kB ± 0%    5.81kB ± 0%     ~     (p=0.310 n=19+20)
DissectWithConversionMultipleValues-8    5.81kB ± 0%    5.81kB ± 0%     ~     (p=0.783 n=20+20)

name                                   old allocs/op  new allocs/op  delta
DissectNoConversionOneValue-8              29.0 ± 0%      29.0 ± 0%     ~     (all equal)
DissectWithConversionOneValue-8            29.0 ± 0%      29.0 ± 0%     ~     (all equal)
DissectNoConversionMultipleValues-8        56.0 ± 0%      56.0 ± 0%     ~     (all equal)
DissectWithConversionMultipleValues-8      56.0 ± 0%      56.0 ± 0%     ~     (all equal)```