@sayden 2021-06-03
/test

Hi @newly12 :slightly_smiling_face: sorry for the delay and thank you for contributing with this. As I told you in the other PR https://github.com/elastic/beats/pull/25256 it is required to have a performance comparison with master written here, in the PR description.

Thanks!