@newly12 2021-06-07
@sayden FYI. I've also [added](https://github.com/elastic/beats/pull/25252/commits/b4f166f7276256d71911d05819ca576f9aaa3834) another benchmark to run dissect in a processor(those DissectProcessor*), it was also executed in current master branch. please have a look at below result.

```
$ benchstat old_master_5.txt new_improvement_5.txt 
name                                                                                   old time/op    new time/op    delta
DissectProcessor/When_all_the_defined_fields_are_captured_by_we_have_remaining_data-8    2.89µs ± 7%    2.78µs ± 3%     ~     (p=0.690 n=5+5)
DissectProcessor/Complex_stack_trace-8                                                   4.35µs ± 6%    4.27µs ± 8%     ~     (p=0.310 n=5+5)
DissectProcessor/fails_when_delimiter_is_not_found_at_the_beginning_of_the_string-8      0.00ns ±47%    0.00ns ±32%     ~     (p=1.000 n=5+5)
DissectProcessor/fails_when_delimiter_is_not_found_after_the_key-8                       0.00ns ±97%    0.00ns ±22%     ~     (p=0.198 n=5+5)
DissectProcessor/simple_dissect-8                                                        1.54µs ± 5%    1.62µs ±10%     ~     (p=0.151 n=5+5)
DissectProcessor/dissect_two_replacement-8                                               2.00µs ± 5%    1.93µs ± 2%     ~     (p=0.056 n=5+5)
DissectProcessor/one_level_dissect_not_end_of_string-8                                   1.91µs ±21%    1.65µs ± 4%     ~     (p=0.056 n=5+5)
DissectProcessor/one_level_dissect-8                                                     1.87µs ±20%    1.55µs ± 1%  -17.07%  (p=0.008 n=5+5)
DissectProcessor/multiple_keys_dissect_end_of_string-8                                   2.19µs ± 3%    1.97µs ± 3%  -10.37%  (p=0.008 n=5+5)
DissectProcessor/multiple_keys_not_end_of_string-8                                       2.07µs ± 2%    1.98µs ± 3%   -4.55%  (p=0.008 n=5+5)
DissectProcessor/simple_ordered-8                                                        1.92µs ± 4%    2.00µs ±16%     ~     (p=1.000 n=5+5)
DissectProcessor/simple_append-8                                                         1.97µs ± 2%    2.00µs ± 6%     ~     (p=0.548 n=5+5)
DissectProcessor/indirect_field-8                                                        1.71µs ± 6%    1.77µs ± 5%     ~     (p=0.222 n=5+5)
DissectProcessor/skip_field-8                                                            1.69µs ± 5%    1.71µs ± 7%     ~     (p=0.952 n=5+5)
DissectProcessor/named_skiped_field_with_indirect-8                                      1.79µs ± 7%    1.71µs ± 3%     ~     (p=0.087 n=5+5)
DissectProcessor/pointer_field_with_indirect-8                                           1.69µs ± 3%    1.73µs ±11%     ~     (p=1.000 n=5+5)
DissectProcessor/missing_fields-8                                                        3.48µs ± 1%    3.54µs ± 3%     ~     (p=0.421 n=5+5)
DissectProcessor/ignore_right_padding-8                                                  2.44µs ± 4%    2.47µs ± 7%     ~     (p=1.000 n=5+5)
DissectProcessor/padding_on_the_last_key_need_a_delimiter-8                              2.44µs ± 0%    2.47µs ± 2%     ~     (p=0.548 n=5+5)
DissectProcessor/ignore_left_padding-8                                                   2.42µs ± 4%    2.49µs ± 6%     ~     (p=0.310 n=5+5)
DissectProcessor/when_the_delimiters_contains_`{`_and_`}`-8                              2.41µs ± 4%    2.32µs ± 2%     ~     (p=0.079 n=5+5)
DissectProcessor/simple_fixed_length-8                                                   2.33µs ± 2%    2.29µs ± 1%     ~     (p=0.056 n=5+5)
DissectProcessor/simple_ordered_and_fixed_length_field-8                                 1.99µs ±11%    1.92µs ± 4%     ~     (p=0.841 n=5+5)
DissectProcessor/simple_padding_and_fixed_length_field-8                                 1.97µs ± 3%    1.92µs ± 3%     ~     (p=0.238 n=5+5)
DissectProcessor/mixed_pointer_and_indirect_and_fixed_length-8                           1.65µs ± 1%    1.66µs ± 4%     ~     (p=1.000 n=5+5)
DissectProcessor/fails_when_there_is_remaining_string_after_the_fixed-length_key-8       0.00ns ±24%    0.00ns ±23%     ~     (p=0.167 n=5+5)
DissectProcessor/fails_when_there_is_no_enough_string_for_the_fixed-length_key-8         0.00ns ±11%    0.00ns ± 8%  +21.61%  (p=0.008 n=5+5)
Dissect/When_all_the_defined_fields_are_captured_by_we_have_remaining_data-8              600ns ± 6%     552ns ± 5%   -7.87%  (p=0.016 n=5+5)
Dissect/Complex_stack_trace-8                                                             818ns ± 0%     838ns ± 9%     ~     (p=0.881 n=5+5)
Dissect/fails_when_delimiter_is_not_found_at_the_beginning_of_the_string-8               0.00ns ±21%    0.00ns ±17%  -31.91%  (p=0.008 n=5+5)
Dissect/fails_when_delimiter_is_not_found_after_the_key-8                                0.00ns ±24%    0.00ns ±36%     ~     (p=0.056 n=5+5)
Dissect/simple_dissect-8                                                                  266ns ± 9%     278ns ± 6%     ~     (p=0.508 n=5+5)
Dissect/dissect_two_replacement-8                                                         314ns ± 2%     351ns ±15%  +11.92%  (p=0.008 n=5+5)
Dissect/one_level_dissect_not_end_of_string-8                                             293ns ± 1%     317ns ± 6%   +8.33%  (p=0.008 n=5+5)
Dissect/one_level_dissect-8                                                               251ns ± 1%     273ns ± 4%   +8.94%  (p=0.008 n=5+5)
Dissect/multiple_keys_dissect_end_of_string-8                                             373ns ± 9%     347ns ± 4%     ~     (p=0.151 n=5+5)
Dissect/multiple_keys_not_end_of_string-8                                                 399ns ± 3%     379ns ± 3%   -5.21%  (p=0.032 n=5+5)
Dissect/simple_ordered-8                                                                  596ns ± 4%     603ns ±11%     ~     (p=1.000 n=5+5)
Dissect/simple_append-8                                                                   580ns ± 6%     600ns ±10%     ~     (p=0.246 n=5+5)
Dissect/indirect_field-8                                                                  376ns ± 6%     385ns ± 6%     ~     (p=0.500 n=5+5)
Dissect/skip_field-8                                                                      355ns ±10%     335ns ± 5%   -5.80%  (p=0.040 n=5+5)
Dissect/named_skiped_field_with_indirect-8                                                395ns ±10%     357ns ± 1%   -9.57%  (p=0.016 n=5+4)
Dissect/pointer_field_with_indirect-8                                                     375ns ± 4%     385ns ± 4%     ~     (p=0.246 n=5+5)
Dissect/missing_fields-8                                                                  662ns ±14%     654ns ± 1%     ~     (p=0.381 n=5+5)
Dissect/ignore_right_padding-8                                                            478ns ± 1%     501ns ± 9%     ~     (p=0.143 n=4+5)
Dissect/padding_on_the_last_key_need_a_delimiter-8                                        585ns ±10%     504ns ± 3%  -13.78%  (p=0.008 n=5+5)
Dissect/ignore_left_padding-8                                                             520ns ±23%     464ns ± 2%     ~     (p=0.056 n=5+5)
Dissect/when_the_delimiters_contains_`{`_and_`}`-8                                        427ns ±15%     407ns ± 3%     ~     (p=0.310 n=5+5)
Dissect/simple_fixed_length-8                                                             433ns ± 8%     389ns ± 1%  -10.21%  (p=0.008 n=5+5)
Dissect/simple_ordered_and_fixed_length_field-8                                           572ns ± 3%     557ns ± 2%     ~     (p=0.095 n=5+5)
Dissect/simple_padding_and_fixed_length_field-8                                           624ns ± 9%     615ns ± 3%     ~     (p=0.651 n=5+5)
Dissect/mixed_pointer_and_indirect_and_fixed_length-8                                     367ns ± 1%     377ns ± 4%     ~     (p=0.103 n=5+5)
Dissect/fails_when_there_is_remaining_string_after_the_fixed-length_key-8                0.00ns ±26%    0.00ns ±22%  -27.23%  (p=0.048 n=5+5)
Dissect/fails_when_there_is_no_enough_string_for_the_fixed-length_key-8                  0.00ns ±33%    0.00ns ±38%     ~     (p=0.056 n=5+5)
Dissect/Regular_expression-8                                                              454ns ± 4%     477ns ± 8%     ~     (p=0.063 n=5+5)
Dissect/Larger_regular_expression-8                                                      3.02µs ± 8%    2.95µs ± 1%     ~     (p=0.690 n=5+5)
Dissect/regular_expression_to_match_end_of_line-8                                         695ns ± 1%     711ns ± 4%     ~     (p=0.095 n=5+5)
DissectNoConversionOneValue-8                                                            5.85µs ± 3%    5.97µs ± 1%     ~     (p=0.095 n=5+5)
DissectWithConversionOneValue-8                                                          6.43µs ± 2%    6.52µs ± 3%     ~     (p=0.421 n=5+5)
DissectNoConversionMultipleValues-8                                                      17.3µs ± 2%    17.9µs ± 4%   +3.63%  (p=0.032 n=5+5)
DissectWithConversionMultipleValues-8                                                    20.7µs ± 4%    20.4µs ± 2%     ~     (p=0.310 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-1-8                                 337ns ± 6%     324ns ± 3%     ~     (p=0.095 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-2-8                                 403ns ± 5%     392ns ± 3%     ~     (p=0.198 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-3-8                                 470ns ± 7%     457ns ± 2%     ~     (p=0.548 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-4-8                                 540ns ± 2%     533ns ± 2%     ~     (p=0.167 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-5-8                                 629ns ± 3%     608ns ± 1%   -3.31%  (p=0.008 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-6-8                                 718ns ± 2%     736ns ± 4%     ~     (p=0.143 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-7-8                                 863ns ± 8%     867ns ±16%     ~     (p=0.548 n=5+5)
DissectComplexStackTraceDegradation/ComplexStackTrace-8-8                                1.23µs ±11%    1.12µs ± 1%   -9.43%  (p=0.040 n=5+5)
StripASCII_4-8                                                                           23.0ns ± 2%    22.5ns ±12%     ~     (p=0.151 n=5+5)
StripASCII_8-8                                                                           26.4ns ±10%    24.3ns ± 3%     ~     (p=0.063 n=5+5)
StripASCII_32-8                                                                          41.4ns ±18%    41.4ns ± 7%     ~     (p=0.635 n=5+5)
StripASCII_128-8                                                                          126ns ±12%     123ns ±16%     ~     (p=1.000 n=5+5)
StripASCII_512-8                                                                          367ns ± 4%     343ns ± 5%   -6.64%  (p=0.032 n=5+5)
StripUTF8_4-8                                                                            58.2ns ± 1%    57.8ns ± 3%     ~     (p=0.730 n=4+5)
StripUTF8_8-8                                                                            93.7ns ± 9%    85.3ns ± 3%     ~     (p=0.056 n=5+5)
StripUTF8_32-8                                                                            204ns ± 2%     196ns ± 1%   -3.92%  (p=0.008 n=5+5)
StripUTF8_128-8                                                                           634ns ± 1%     616ns ± 1%   -2.87%  (p=0.008 n=5+5)
StripUTF8_512-8                                                                          2.57µs ± 6%    2.34µs ± 2%   -9.06%  (p=0.008 n=5+5)

name                                                                                   old alloc/op   new alloc/op   delta
DissectProcessor/When_all_the_defined_fields_are_captured_by_we_have_remaining_data-8    1.97kB ± 0%    1.97kB ± 0%     ~     (all equal)
DissectProcessor/Complex_stack_trace-8                                                   2.16kB ± 0%    2.16kB ± 0%     ~     (all equal)
DissectProcessor/fails_when_delimiter_is_not_found_at_the_beginning_of_the_string-8       0.00B          0.00B          ~     (all equal)
DissectProcessor/fails_when_delimiter_is_not_found_after_the_key-8                        0.00B          0.00B          ~     (all equal)
DissectProcessor/simple_dissect-8                                                        1.81kB ± 0%    1.81kB ± 0%     ~     (all equal)
DissectProcessor/dissect_two_replacement-8                                               1.86kB ± 0%    1.86kB ± 0%     ~     (all equal)
DissectProcessor/one_level_dissect_not_end_of_string-8                                   1.81kB ± 0%    1.81kB ± 0%     ~     (all equal)
DissectProcessor/one_level_dissect-8                                                     1.81kB ± 0%    1.81kB ± 0%     ~     (all equal)
DissectProcessor/multiple_keys_dissect_end_of_string-8                                   1.86kB ± 0%    1.86kB ± 0%     ~     (all equal)
DissectProcessor/multiple_keys_not_end_of_string-8                                       1.86kB ± 0%    1.86kB ± 0%     ~     (all equal)
DissectProcessor/simple_ordered-8                                                        1.84kB ± 0%    1.84kB ± 0%     ~     (all equal)
DissectProcessor/simple_append-8                                                         1.84kB ± 0%    1.84kB ± 0%     ~     (all equal)
DissectProcessor/indirect_field-8                                                        1.82kB ± 0%    1.82kB ± 0%     ~     (all equal)
DissectProcessor/skip_field-8                                                            1.82kB ± 0%    1.82kB ± 0%     ~     (all equal)
DissectProcessor/named_skiped_field_with_indirect-8                                      1.82kB ± 0%    1.82kB ± 0%     ~     (all equal)
DissectProcessor/pointer_field_with_indirect-8                                           1.82kB ± 0%    1.82kB ± 0%     ~     (all equal)
DissectProcessor/missing_fields-8                                                        2.02kB ± 0%    2.02kB ± 0%     ~     (all equal)
DissectProcessor/ignore_right_padding-8                                                  1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
DissectProcessor/padding_on_the_last_key_need_a_delimiter-8                              1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
DissectProcessor/ignore_left_padding-8                                                   1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
DissectProcessor/when_the_delimiters_contains_`{`_and_`}`-8                              1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
DissectProcessor/simple_fixed_length-8                                                   1.90kB ± 0%    1.90kB ± 0%     ~     (all equal)
DissectProcessor/simple_ordered_and_fixed_length_field-8                                 1.84kB ± 0%    1.84kB ± 0%     ~     (all equal)
DissectProcessor/simple_padding_and_fixed_length_field-8                                 1.84kB ± 0%    1.84kB ± 0%     ~     (all equal)
DissectProcessor/mixed_pointer_and_indirect_and_fixed_length-8                           1.82kB ± 0%    1.82kB ± 0%     ~     (all equal)
DissectProcessor/fails_when_there_is_remaining_string_after_the_fixed-length_key-8        0.00B          0.00B          ~     (all equal)
DissectProcessor/fails_when_there_is_no_enough_string_for_the_fixed-length_key-8          0.00B          0.00B          ~     (all equal)
Dissect/When_all_the_defined_fields_are_captured_by_we_have_remaining_data-8               400B ± 0%      400B ± 0%     ~     (all equal)
Dissect/Complex_stack_trace-8                                                              464B ± 0%      464B ± 0%     ~     (all equal)
Dissect/fails_when_delimiter_is_not_found_at_the_beginning_of_the_string-8                0.00B          0.00B          ~     (all equal)
Dissect/fails_when_delimiter_is_not_found_after_the_key-8                                 0.00B          0.00B          ~     (all equal)
Dissect/simple_dissect-8                                                                   352B ± 0%      352B ± 0%     ~     (all equal)
Dissect/dissect_two_replacement-8                                                          368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/one_level_dissect_not_end_of_string-8                                              352B ± 0%      352B ± 0%     ~     (all equal)
Dissect/one_level_dissect-8                                                                352B ± 0%      352B ± 0%     ~     (all equal)
Dissect/multiple_keys_dissect_end_of_string-8                                              368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/multiple_keys_not_end_of_string-8                                                  368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/simple_ordered-8                                                                   392B ± 0%      392B ± 0%     ~     (all equal)
Dissect/simple_append-8                                                                    392B ± 0%      392B ± 0%     ~     (all equal)
Dissect/indirect_field-8                                                                   368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/skip_field-8                                                                       368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/named_skiped_field_with_indirect-8                                                 368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/pointer_field_with_indirect-8                                                      368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/missing_fields-8                                                                   432B ± 0%      432B ± 0%     ~     (all equal)
Dissect/ignore_right_padding-8                                                             384B ± 0%      384B ± 0%     ~     (all equal)
Dissect/padding_on_the_last_key_need_a_delimiter-8                                         384B ± 0%      384B ± 0%     ~     (all equal)
Dissect/ignore_left_padding-8                                                              384B ± 0%      384B ± 0%     ~     (all equal)
Dissect/when_the_delimiters_contains_`{`_and_`}`-8                                         384B ± 0%      384B ± 0%     ~     (all equal)
Dissect/simple_fixed_length-8                                                              384B ± 0%      384B ± 0%     ~     (all equal)
Dissect/simple_ordered_and_fixed_length_field-8                                            392B ± 0%      392B ± 0%     ~     (all equal)
Dissect/simple_padding_and_fixed_length_field-8                                            392B ± 0%      392B ± 0%     ~     (all equal)
Dissect/mixed_pointer_and_indirect_and_fixed_length-8                                      368B ± 0%      368B ± 0%     ~     (all equal)
Dissect/fails_when_there_is_remaining_string_after_the_fixed-length_key-8                 0.00B          0.00B          ~     (all equal)
Dissect/fails_when_there_is_no_enough_string_for_the_fixed-length_key-8                   0.00B          0.00B          ~     (all equal)
Dissect/Regular_expression-8                                                              0.00B          0.00B          ~     (all equal)
Dissect/Larger_regular_expression-8                                                        532B ± 0%      532B ± 0%     ~     (p=0.444 n=5+5)
Dissect/regular_expression_to_match_end_of_line-8                                          307B ± 0%      307B ± 0%     ~     (p=0.556 n=5+4)
DissectNoConversionOneValue-8                                                            2.64kB ± 0%    2.64kB ± 0%     ~     (p=0.278 n=5+5)
…